\name{cornell0}
\alias{cornell0}
\docType{data}
\title{
Motor Octane Number Measurements
}
\description{
Motor octane number measurements in different mixtures.
}
\usage{data("cornell0")}
\format{
  A data frame with 12 observations on the following 8 variables.
All are numeric.
\tabular{rll}{
    [,1] \tab Distillation  \tab Direct distillation\cr
    [,2]  \tab Reformat  \tab Reformat \cr
        [,3]  \tab NaphthaT  \tab    Naphtha thermal cracked
	\cr
        [,4]  \tab NaphthaC   \tab Naphtha catalytic
	cracked\cr
	 [,5] \tab Polymer   \tab
	 Polymer\cr
	 [,6] \tab Alkylat  \tab
	 Alkylat\cr
	 [, 7] \tab Gasoline   \tab 
	 Natural gasoline\cr
	 [, 8]  \tab Y   \tab 
     Octane number\cr
   }
}

\source{
 Kettaneh-Wold N. (1992)
    Analysis of mixture data with partial least squares.
    Chemometrics and Intellignet Laboratory Systems,
    vol. 14, pp. 57-69.
}
\references{
  \itemize{
    \item Tenenhaus M., Gauchi J.-P.  and Ménardo C. (1995) Régression PLS
  et applications. Revue de Statistique Appliquée, vol. 53(1), pp. 7-63.

  \item Tenenhaus M. (1998) La régression PLS. Théorie et
  pratique. Ed. TECHNIP. Paris.
  }
}
\keyword{datasets}
