\name{cornell1}
\alias{cornell1}
\docType{data}
\title{
Motor Octane Number Measurements
}
\description{
This dataset is the \code{\link{cornell0}} dataset where missing values have
  been introduced in the first 7 variables.
}
\usage{data("cornell1")}
\source{
    Kettaneh-Wold N. (1992)
    Analysis of mixture data with partial least squares.
    Chemometrics and Intellignet Laboratory Systems,
    vol. 14, pp. 57-69.

}
\references{
  \itemize{
    \item Tenenhaus M., Gauchi J.-P.  and Ménardo C. (1995) Régression PLS
  et applications. Revue de Statistique Appliquée, vol. 53(1), pp. 7-63.

  \item Tenenhaus M. (1998) La régression PLS. Théorie et
  pratique. Ed. TECHNIP. Paris.
  }
}
\keyword{datasets}
