\name{XY180}
\alias{XY180}
\docType{data}
\title{
Transformed Aircraft InfraRed Signature
}
\description{
This dataset is the result of monomial calculation on the
 dataset \code{\link{X18Y2}}
(Aircraft InfraRed Signature).
  The monomials are described in the vignette of the package.
}
\usage{data("XY180")}
\format{
  A data frame with 180 observations and 160 variables, labelled
  "V1", ... "V160".
  The 158 first ones are monomial values. The  last two variables
  are response values.
}
\source{
Lefebvre, S. ONERA, Palaiseau, France.
}
\references{
Gauchi, J.-P. 2015.
      A practical method of global sensitivity analysis
under constraints.
Rapport technique 2015-1.
INRA, UR1404, F-78350 Jouy-en-Josas, France.
}
\keyword{datasets}
